+++ modules/shared/core/src/sap/commerce/toolset/Plugin.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -29,6 +29,7 @@ enum class Plugin(val id: String, val url: String? = null, val dependencies: Set
 
     HYBRIS(HybrisConstants.PLUGIN_ID, url = "https://plugins.jetbrains.com/plugin/12867-sap-commerce-developers-toolset"),
     ULTIMATE("com.intellij.modules.ultimate"),
+    AngularBackend("intellij.angular.backend"),
     JREBEL("JRebelPlugin", url = "https://plugins.jetbrains.com/plugin/4441-jrebel-and-xrebel"),
     ANT_SUPPORT("AntSupport", url = "https://plugins.jetbrains.com/plugin/23025-ant"),
     MAVEN("org.jetbrains.idea.maven"),
@@ -36,7 +37,7 @@ enum class Plugin(val id: String, val url: String? = null, val dependencies: Set
     GROOVY("org.intellij.groovy", url = "https://plugins.jetbrains.com/plugin/1524-groovy"),
     GRADLE("com.intellij.gradle"),
     ECLIPSE("org.jetbrains.idea.eclipse"),
-    ANGULAR("AngularJS", url = "https://plugins.jetbrains.com/plugin/6971-angular"),
+    ANGULAR("AngularJS", url = "https://plugins.jetbrains.com/plugin/6971-angular", dependencies = setOf(AngularBackend)),
     DATABASE("com.intellij.database", url = "https://plugins.jetbrains.com/plugin/10925-database-tools-and-sql-for-webstorm", dependencies = setOf(ULTIMATE)),
     DIAGRAM("com.intellij.diagram"),
     PROPERTIES("com.intellij.properties", url = "https://plugins.jetbrains.com/plugin/11594-properties"),